backend/pyproject.toml
[project]
name = "locavox"
version = "0.1.0"
description = ""
authors = [
    {name = "Jean-Baptiste Poullet",email = "jeanbaptistepoullet@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0,<1.0.0",
    "uvicorn>=0.34.0",
    "pydantic[email] (>=2.10.6,<3.0.0)",
    "openai (>=1.64.0,<2.0.0)",
    "lancedb (>=0.20.0,<0.21.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "python-jose[cryptography] (>=3.4.0,<4.0.0)",
    "passlib[bcrypt]>=1.7.4,<1.8.0",
    "bcrypt>=4.0.0,<4.1.0",
    "python-multipart (>=0.0.20,<0.0.21)",
    "sqlalchemy (>=2.0.38,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
pytest-asyncio = "^0.25.3"
ruff = "^0.9.9"


[tool.poetry.group.embeddings.dependencies]
sentence-transformers = "^3.4.1"


[tool.poetry.group.database.dependencies]
asyncpg = "^0.30.0"


[tool.poetry.group.database-dev.dependencies]
aiosqlite = "^0.21.0"

